# api_gateway/api/docs/swagger.yaml
definitions:
  http.Response:
    properties:
      data: {}
      description:
        type: string
      status:
        type: string
    type: object
...
  schedule_service.CreateAssignStudent:
    properties:
      event_id:
        type: string
      student_id:
        type: string
    type: object
...
  schedule_service.GetListAssignStudentResponse:
    properties:
      AssignStudents:
        items:
          $ref: '#/definitions/schedule_service.AssignStudent'
        type: array
      count:
        type: integer
    type: object
...
  user_service.Administrator:
    properties:
      branch_id:
        type: string
      created_at:
        type: string
      full_name:
        type: string
      id:
        type: string
      ielts_score:
        type: string
      login:
        type: string
      password:
        type: string
      phone:
        type: string
      role_id:
        type: string
      salary:
        type: number
      updated_at:
        type: string
    type: object
  user_service.AdministratorPanelResponse:
    properties:
      Payments:
        items:
          $ref: '#/definitions/user_service.Payment'
        type: array
      Students:
        items:
          $ref: '#/definitions/user_service.Student'
        type: array
      SupportTeachers:
        items:
          $ref: '#/definitions/user_service.SupportTeacher'
        type: array
    type: object
...
  user_service.CreateAdministrator:
    properties:
      branch_id:
        type: string
      full_name:
        type: string
      ielts_score:
        type: string
      password:
        type: string
      phone:
        type: string
      salary:
        type: number
    type: object
  user_service.CreateBranch:
    properties:
      name:
        type: string
    type: object
...
  user_service.GetListAdministratorResponse:
    properties:
      Administrators:
        items:
          $ref: '#/definitions/user_service.Administrator'
        type: array
      count:
        type: integer
    type: object
...
  user_service.UpdateAdministrator:
    properties:
      branch_id:
        type: string
      full_name:
        type: string
      id:
        type: string
      ielts_score:
        type: string
      password:
        type: string
      phone:
        type: string
      salary:
        type: number
    type: object
...
info:
  contact: {}
paths:
  /administrator:
    get:
      consumes:
      - application/json
      description: Get Administrators List
      operationId: get_Administrator_list
      parameters:
      - description: role_id
        in: query
        name: role_id
        required: true
        type: string
      - description: offset
        in: query
        name: offset
        type: integer
      - description: limit
        in: query
        name: limit
        type: integer
      - description: search
        in: query
        name: search
        type: string
      - default: a1924766-a9ee-11ed-afa1-0242ac120001
        description: Platform-Id
        in: header
        name: Platform-Id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: GetAllAdministratorResponseBody
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/user_service.GetListAdministratorResponse'
              type: object
        "400":
          description: Invalid Argument
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: Get Administrators List
      tags:
      - Administrator
    post:
      consumes:
      - application/json
      description: Create Administrator
      operationId: create_Administrator
      parameters:
      - description: role_id
        in: query
        name: role_id
        required: true
        type: string
      - description: CreateAdministratorBody
        in: body
        name: profile
        required: true
        schema:
          $ref: '#/definitions/user_service.CreateAdministrator'
      produces:
      - application/json
      responses:
        "200":
          description: GetAdministratorBody
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/user_service.Administrator'
              type: object
        "400":
          description: Invalid Argument
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Create Administrator
      tags:
      - Administrator
  /administrator/{id}:
    delete:
      consumes:
      - application/json
      description: Delete Administrator
      operationId: delete_Administrator
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      - description: role_id
        in: query
        name: role_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Administrator data
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Delete Administrator
      tags:
      - Administrator
    get:
      consumes:
      - application/json
      description: Get Administrator  By ID
      operationId: get_Administrator_by_id
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      - description: role_id
        in: query
        name: role_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: AdministratorBody
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/user_service.Administrator'
              type: object
        "400":
          description: Invalid Argument
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Get Administrator  By ID
      tags:
      - Administrator
    put:
      consumes:
      - application/json
      description: Update Administrator
      operationId: update_Administrator
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      - description: role_id
        in: query
        name: role_id
        required: true
        type: string
      - description: UpdateAdministrator
        in: body
        name: profile
        required: true
        schema:
          $ref: '#/definitions/user_service.UpdateAdministrator'
      produces:
      - application/json
      responses:
        "200":
          description: Administrator data
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/user_service.Administrator'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Update Administrator
      tags:
      - Administrator
  /adminstrator-panel/{id}:
    get:
      consumes:
      - application/json
      description: Panel Administrator
      operationId: adminstrator_panel
      parameters:
      - description: id
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: AdministratorPanelResponse
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/user_service.AdministratorPanelResponse'
              type: object
        "400":
          description: Invalid Argument
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: adminstrator_panel
      tags:
      - Administrator
  /assign-student:
    get:
      consumes:
      - application/json
      description: Get AssignStudents List
      operationId: get_assignstudent_list
      parameters:
      - description: offset
        in: query
        name: offset
        type: integer
      - description: limit
        in: query
        name: limit
        type: integer
      - description: search
        in: query
        name: search
        type: string
      - default: a1924766-a9ee-11ed-afa1-0242ac120001
        description: Platform-Id
        in: header
        name: Platform-Id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: GetAllAssignStudentResponseBody
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/schedule_service.GetListAssignStudentResponse'
              type: object
        "400":
          description: Invalid Argument
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: Get AssignStudents List
      tags:
      - AssignStudent
    post:
      consumes:
      - application/json
      description: Create AssignStudent
      operationId: create_assignstudent
      parameters:
      - description: CreateAssignStudentBody
        in: body
        name: profile
        required: true
        schema:
          $ref: '#/definitions/schedule_service.CreateAssignStudent'
      produces:
      - application/json
      responses:
        "200":
          description: GetAssignStudentBody
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/schedule_service.AssignStudent'
              type: object
        "400":
          description: Invalid Argument
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Create AssignStudent
      tags:
      - AssignStudent
  /assign-student/{id}:
    delete:
      consumes:
      - application/json
      description: Delete AssignStudent
      operationId: delete_assignstudent
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: AssignStudent data
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Delete AssignStudent
      tags:
      - AssignStudent
    get:
      consumes:
      - application/json
      description: Get AssignStudent  By ID
      operationId: get_assignstudent_by_id
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: AssignStudentBody
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/schedule_service.AssignStudent'
              type: object
        "400":
          description: Invalid Argument
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Get AssignStudent  By ID
      tags:
      - AssignStudent
    put:
      consumes:
      - application/json
      description: Update AssignStudent
      operationId: update_assignstudent
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      - description: UpdateAssignStudent
        in: body
        name: profile
        required: true
        schema:
          $ref: '#/definitions/schedule_service.UpdateAssignStudent'
      produces:
      - application/json
      responses:
        "200":
          description: AssignStudent data
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/schedule_service.AssignStudent'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Update AssignStudent
      tags:
      - AssignStudent
  /branch:
    get:
      consumes:
      - application/json
      description: Get Branchs List
      operationId: get_Branch_list
      parameters:
      - description: role_id
        in: query
        name: role_id
        required: true
        type: string
      - description: offset
        in: query
        name: offset
        type: integer
      - description: limit
        in: query
        name: limit
        type: integer
      - description: search
        in: query
        name: search
        type: string
      - default: a1924766-a9ee-11ed-afa1-0242ac120001
        description: Platform-Id
        in: header
        name: Platform-Id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: GetAllBranchResponseBody
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/user_service.GetListBranchResponse'
              type: object
        "400":
          description: Invalid Argument
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: Get Branchs List
      tags:
      - Branch
    post:
      consumes:
      - application/json
      description: Create Branch
      operationId: create_Branch
      parameters:
      - description: role_id
        in: query
        name: role_id
        required: true
        type: string
      - description: CreateBranchBody
        in: body
        name: profile
        required: true
        schema:
          $ref: '#/definitions/user_service.CreateBranch'
      produces:
      - application/json
      responses:
        "200":
          description: GetBranchBody
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/user_service.Branch'
              type: object
        "400":
          description: Invalid Argument
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Create Branch
      tags:
      - Branch
  /branch/{id}:
    delete:
      consumes:
      - application/json
      description: Delete Branch
      operationId: delete_Branch
      parameters:
      - description: role_id
        in: query
        name: role_id
        required: true
        type: string
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Branch data
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Delete Branch
      tags:
      - Branch
    get:
      consumes:
      - application/json
      description: Get Branch  By ID
      operationId: get_Branch_by_id
      parameters:
      - description: role_id
        in: query
        name: role_id
        required: true
        type: string
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: BranchBody
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/user_service.Branch'
              type: object
        "400":
          description: Invalid Argument
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Get Branch  By ID
      tags:
      - Branch
    put:
      consumes:
      - application/json
      description: Update Branch
      operationId: update_Branch
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      - description: role_id
        in: query
        name: role_id
        required: true
        type: string
      - description: UpdateBranch
        in: body
        name: profile
        required: true
        schema:
          $ref: '#/definitions/user_service.UpdateBranch'
      produces:
      - application/json
      responses:
        "200":
          description: Branch data
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/user_service.Branch'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Update Branch
      tags:
      - Branch
  /dotask:
    get:
      consumes:
      - application/json
      description: Get DoTasks List
      operationId: get_dotask_list
      parameters:
      - description: offset
        in: query
        name: offset
        type: integer
      - description: limit
        in: query
        name: limit
        type: integer
      - description: search
        in: query
        name: search
        type: string
      - default: a1924766-a9ee-11ed-afa1-0242ac120001
        description: Platform-Id
        in: header
        name: Platform-Id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: GetAllDoTaskResponseBody
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/schedule_service.GetListDoTaskResponse'
              type: object
        "400":
          description: Invalid Argument
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: Get DoTasks List
      tags:
      - DoTask
    post:
      consumes:
      - application/json
      description: Create DoTask
      operationId: create_dotask
      parameters:
      - description: CreateDoTaskBody
        in: body
        name: profile
        required: true
        schema:
          $ref: '#/definitions/schedule_service.CreateDoTask'
      produces:
      - application/json
      responses:
        "200":
          description: GetDoTaskBody
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/schedule_service.DoTask'
              type: object
        "400":
          description: Invalid Argument
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Create DoTask
      tags:
      - DoTask
...
  /event:
    get:
      consumes:
      - application/json
      description: Get Events List
      operationId: get_event_list
      parameters:
      - description: role_id
        in: query
        name: role_id
        required: true
        type: string
      - description: offset
        in: query
        name: offset
        type: integer
      - description: limit
        in: query
        name: limit
        type: integer
      - description: search
        in: query
        name: search
        type: string
      - default: a1924766-a9ee-11ed-afa1-0242ac120001
        description: Platform-Id
        in: header
        name: Platform-Id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: GetAllEventResponseBody
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/schedule_service.GetListEventResponse'
              type: object
        "400":
          description: Invalid Argument
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: Get Events List
      tags:
      - Event
    post:
      consumes:
      - application/json
      description: Create Event
      operationId: create_event
      parameters:
      - description: role_id
        in: query
        name: role_id
        required: true
        type: string
      - description: CreateEventBody
        in: body
        name: profile
        required: true
        schema:
          $ref: '#/definitions/schedule_service.CreateEvent'
      produces:
      - application/json
      responses:
        "200":
          description: GetEventBody
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/schedule_service.Event'
              type: object
        "400":
          description: Invalid Argument
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Create Event
      tags:
      - Event
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"